Feat: Add the component server entry
