+++ docs/src/pages/test-ssr/ladder.astro
@@ -3,7 +3,7 @@ import Body from '@layouts/Body.astro';
 import '@css/layers.css';
 import '@css/fonts.css';
 
-import { renderToString } from '@semantic-ui/component';
+import { renderToString } from '@semantic-ui/component/server';
 import {
   Step1, Step2, Step3, Step4, Step5, Step6, Step7, Step8,
   Step9, Step10, Step11, Step12, Step13, Step14, Step15, Step16,
