+++ docs/src/pages/test-ssr/vanilla.astro
@@ -9,7 +9,7 @@ import '@css/layers.css';
 import '@css/fonts.css';
 import '@semantic-ui/core/icon/sets/lucide.css';
 
-import { renderToString } from '@semantic-ui/component';
+import { renderToString } from '@semantic-ui/component/server';
 import { NavMenu } from '@semantic-ui/core';
 
 const menu = [
