+++ packages/component/src/engines/native/register.js
@@ -1,7 +1,8 @@
-import { registerEngine, Renderer, ServerRenderer } from '@semantic-ui/renderer';
+import { registerEngine, Renderer } from '@semantic-ui/renderer';
 import { createComponent } from './factory.js';
 
-const NativeEngine = { renderer: Renderer, serverRenderer: ServerRenderer, factory: createComponent };
+// the server renderer joins on the server entry, so the root stays browser-sized
+const NativeEngine = { renderer: Renderer, factory: createComponent };
 registerEngine('native', NativeEngine);
 
 export { NativeEngine };
