+++ packages/component/types/index.d.ts
@@ -5,10 +5,7 @@ export { getComponent, hasComponent, registerComponent } from './component-regis
 export { defineComponent } from './define-component.js';
 export { WebComponentBase } from './engines/native/base.js';
 export { NativeEngine } from './engines/native/register.js';
-export { expandCustomElements } from './expand-custom-elements.js';
 export { setRecovery, setStackCapture, setTracing } from './helpers.js';
-export { renderToStaticMarkup } from './render-to-static-markup.js';
-export { renderToString } from './render-to-string.js';
 
 export type {
   ComponentConstructor,
@@ -19,6 +16,3 @@ export type {
   KeyCallParams,
 } from './define-component.js';
 export type { ComponentConfig, GetPropertiesOptions, PropertyConfig } from './engines/native/base.js';
-export type { ExpandCustomElementsOptions } from './expand-custom-elements.js';
-export type { RenderToStaticMarkupOptions, StaticMarkup } from './render-to-static-markup.js';
-export type { RenderToStringOptions } from './render-to-string.js';
