+++ packages/component/types/server.d.ts
@@ -1,0 +1,9 @@
+export * from './index.js';
+
+export { expandCustomElements } from './expand-custom-elements.js';
+export { renderToStaticMarkup } from './render-to-static-markup.js';
+export { renderToString } from './render-to-string.js';
+
+export type { ExpandCustomElementsOptions } from './expand-custom-elements.js';
+export type { RenderToStaticMarkupOptions, StaticMarkup } from './render-to-static-markup.js';
+export type { RenderToStringOptions } from './render-to-string.js';
